refactor: move classes in safehill client to respective package
