+++ src/main/kotlin/com/safehill/safehillclient/data/factory/RepositoriesFactory.kt
@@ -1,19 +1,15 @@
-package com.safehill.safehillclient.data.threads.factory
+package com.safehill.safehillclient.data.factory
 
 import com.safehill.kclient.controllers.module.ControllersModule
 import com.safehill.kclient.models.users.UserProvider
-import com.safehill.kclient.network.ServerProxy
 import com.safehill.safehillclient.backgroundsync.BackgroundTasksRegistry
 import com.safehill.safehillclient.backgroundsync.ClientOptions
 import com.safehill.safehillclient.backgroundsync.NetworkModule
 import com.safehill.safehillclient.data.authorization.UserAuthorizationRepository
-import com.safehill.safehillclient.data.factory.MessageInteractorFactory
 import com.safehill.safehillclient.data.threads.ThreadsRepository
-import com.safehill.safehillclient.data.threads.interactor.ThreadStateInteractor
-import com.safehill.safehillclient.data.threads.model.MutableThreadState
+import com.safehill.safehillclient.data.threads.factory.ThreadStateInteractorFactory
 import com.safehill.safehillclient.data.threads.registry.ThreadStateRegistry
 import com.safehill.safehillclient.data.user_discovery.UserDiscoveryRepository
-import kotlinx.coroutines.CoroutineScope
 
 class RepositoriesFactory(
     private val networkModule: NetworkModule,
@@ -55,34 +51,6 @@ class RepositoriesFactory(
         return UserDiscoveryRepository(
             serverProxy = networkModule.serverProxy,
             sdkDispatchers = clientOptions.sdkDispatchers
-        )
-    }
-}
-
-class ThreadStateInteractorFactory(
-    private val serverProxy: ServerProxy,
-    private val controllersModule: ControllersModule,
-    private val userProvider: UserProvider,
-    private val clientOptions: ClientOptions
-) {
-    fun create(
-        threadID: String,
-        scope: CoroutineScope,
-        mutableThreadState: MutableThreadState
-    ): ThreadStateInteractor {
-        return ThreadStateInteractor(
-            threadId = threadID,
-            scope = scope,
-            mutableThreadState = mutableThreadState,
-            serverProxy = serverProxy,
-            userController = controllersModule.userController,
-            messageInteractorFactory = MessageInteractorFactory(
-                interactionController = controllersModule.interactionController,
-                userScope = scope,
-                userProvider = userProvider
-            ),
-            userProvider = userProvider,
-            safehillLogger = clientOptions.safehillLogger
         )
     }
 }
