+++ src/main/kotlin/com/safehill/safehillclient/data/message/factory/MessageInteractorFactory.kt
@@ -1,11 +1,11 @@
-package com.safehill.safehillclient.data.factory
+package com.safehill.safehillclient.data.message.factory
 
 import com.safehill.kclient.controllers.UserInteractionController
 import com.safehill.kclient.models.interactions.InteractionAnchor
 import com.safehill.kclient.models.users.UserProvider
-import com.safehill.safehillclient.data.threads.interactor.MessageInteractorImpl
-import com.safehill.safehillclient.data.threads.model.MutableMessagesContainer
-import com.safehill.safehillclient.module.sdk.UserScope
+import com.safehill.safehillclient.data.message.model.MutableMessagesContainer
+import com.safehill.safehillclient.data.message.interactor.MessageInteractorImpl
+import com.safehill.safehillclient.module.client.UserScope
 
 class MessageInteractorFactory(
     private val interactionController: UserInteractionController,
