+++ src/main/kotlin/com/safehill/safehillclient/data/threads/interactor/ThreadStateInteractor.kt
@@ -10,12 +10,13 @@ import com.safehill.kclient.models.users.UserProvider
 import com.safehill.kclient.models.users.getOrNull
 import com.safehill.kclient.network.ServerProxy
 import com.safehill.kclient.util.safeApiCall
-import com.safehill.safehillclient.data.factory.MessageInteractorFactory
-import com.safehill.safehillclient.data.threads.model.Message
-import com.safehill.safehillclient.data.threads.model.MessageStatus
-import com.safehill.safehillclient.data.threads.model.MessageType
+import com.safehill.safehillclient.data.message.factory.MessageInteractorFactory
+import com.safehill.safehillclient.data.message.interactor.MessageInteractor
+import com.safehill.safehillclient.data.message.model.Message
+import com.safehill.safehillclient.data.message.model.MessageStatus
+import com.safehill.safehillclient.data.message.model.MessageType
 import com.safehill.safehillclient.data.threads.model.MutableThreadState
-import com.safehill.safehillclient.model.toAppUser
+import com.safehill.safehillclient.model.user.toAppUser
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import java.time.Instant
