+++ src/main/kotlin/com/safehill/safehillclient/error/ClientErrors.kt
