+++ src/main/kotlin/com/safehill/safehillclient/manager/dependencies/SdkRepositories.kt
@@ -1,4 +1,4 @@
-package com.safehill.safehillclient.dependencies
+package com.safehill.safehillclient.manager.dependencies
 
 import com.safehill.kclient.controllers.module.ControllersModule
 import com.safehill.kclient.models.users.LocalUser
@@ -7,10 +7,10 @@ import com.safehill.safehillclient.backgroundsync.BackgroundTasksRegistry
 import com.safehill.safehillclient.backgroundsync.ClientOptions
 import com.safehill.safehillclient.backgroundsync.NetworkModule
 import com.safehill.safehillclient.data.authorization.UserAuthorizationRepository
+import com.safehill.safehillclient.data.factory.RepositoriesFactory
 import com.safehill.safehillclient.data.threads.ThreadsRepository
-import com.safehill.safehillclient.data.threads.factory.RepositoriesFactory
 import com.safehill.safehillclient.data.user_discovery.UserDiscoveryRepository
-import com.safehill.safehillclient.module.sdk.ClientModule
+import com.safehill.safehillclient.module.client.ClientModule
 
 class SdkRepositories private constructor(
     backgroundTaskRegistry: BackgroundTasksRegistry,
