+++ src/main/kotlin/com/safehill/safehillclient/manager/dependencies/UserObserver.kt
@@ -1,4 +1,4 @@
-package com.safehill.safehillclient.dependencies
+package com.safehill.safehillclient.manager.dependencies
 
 import com.safehill.kclient.models.users.LocalUser
