+++ src/main/kotlin/com/safehill/safehillclient/module/client/ClientModule.kt
@@ -1,4 +1,4 @@
-package com.safehill.safehillclient.module.sdk
+package com.safehill.safehillclient.module.client
 
 import com.safehill.kclient.controllers.module.ControllersModule
 import com.safehill.kclient.errors.LocalUserError
@@ -6,13 +6,11 @@ import com.safehill.kclient.models.users.LocalUser
 import com.safehill.kclient.util.Provider
 import com.safehill.safehillclient.backgroundsync.ClientOptions
 import com.safehill.safehillclient.backgroundsync.SafehillBackgroundTasksRegistryFactory
-import com.safehill.safehillclient.backgroundsync.SafehillSyncManager
-import com.safehill.safehillclient.dependencies.SdkRepositories
-import com.safehill.safehillclient.dependencies.UserObserver
+import com.safehill.safehillclient.manager.dependencies.UserObserver
 import com.safehill.safehillclient.module.asset.AssetModule
-import com.safehill.safehillclient.platform.PlatformModule
-import com.safehill.safehillclient.platform.UserModule
-import com.safehill.safehillclient.serverproxy.NetworkModuleFactory
+import com.safehill.safehillclient.module.platform.PlatformModule
+import com.safehill.safehillclient.module.platform.UserModule
+import com.safehill.safehillclient.factory.NetworkModuleFactory
 import kotlinx.coroutines.CoroutineScope
 import java.util.Collections
 import java.util.concurrent.atomic.AtomicReference
